+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### 0.0.3
 * When the build fails, show the output in Bar Output.
+* Ask before creating bar.conf.json
+* Autoload the extension if `.vscode/bar.conf.json` exists.
+* Can't initialize the extension more than once. (bug fix)
+* Minor improvements.
+* Added more bugs.
 
 ### 0.0.2
 * Updated README.md